调整测试用例.
