+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/h2/optimisticLocker/entity/IntVersionUser.java
@@ -1,14 +1,12 @@
-package com.baomidou.mybatisplus.test.plugins.optimisticLocker.entity;
+package com.baomidou.mybatisplus.test.h2.optimisticLocker.entity;
 
 import java.io.Serializable;
-import java.util.Date;
 
 import com.baomidou.mybatisplus.annotations.TableField;
 import com.baomidou.mybatisplus.annotations.TableName;
-import com.baomidou.mybatisplus.annotations.Version;
 
-@TableName("time_version_user")
-public class DateVersionUser implements Serializable {
+@TableName("version_user")
+public class IntVersionUser extends IntVersionFather implements Serializable {
 
     @TableField(exist = false)
     private static final long serialVersionUID = 1L;
@@ -17,8 +15,7 @@ public class DateVersionUser implements Serializable {
 
     private String name;
 
-    @Version
-    private Date version;
+    private Integer age;
 
     public Long getId() {
         return id;
@@ -36,12 +33,12 @@ public class DateVersionUser implements Serializable {
         this.name = name;
     }
 
-    public Date getVersion() {
-        return version;
+    public Integer getAge() {
+        return age;
     }
 
-    public void setVersion(Date version) {
-        this.version = version;
+    public void setAge(Integer age) {
+        this.age = age;
     }
 
 }
