+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/h2/optimisticLocker/mapper/LongVersionUserMapper.java
@@ -1,0 +1,8 @@
+package com.baomidou.mybatisplus.test.h2.optimisticLocker.mapper;
+
+import com.baomidou.mybatisplus.mapper.BaseMapper;
+import com.baomidou.mybatisplus.test.h2.optimisticLocker.entity.LongVersionUser;
+
+public interface LongVersionUserMapper extends BaseMapper<LongVersionUser> {
+
+}
