+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/h2/optimisticLocker/mapper/TimestampVersionUserMapper.java
@@ -1,0 +1,8 @@
+package com.baomidou.mybatisplus.test.h2.optimisticLocker.mapper;
+
+import com.baomidou.mybatisplus.mapper.BaseMapper;
+import com.baomidou.mybatisplus.test.h2.optimisticLocker.entity.TimestampVersionUser;
+
+public interface TimestampVersionUserMapper extends BaseMapper<TimestampVersionUser> {
+
+}
