+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/mysql/CrudTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.baomidou.mybatisplus.test;
+package com.baomidou.mybatisplus.test.mysql;
 
 import java.io.InputStream;
