+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/mysql/paginationInterceptor/mapper/PageUserMapper.java
@@ -1,0 +1,8 @@
+package com.baomidou.mybatisplus.test.mysql.paginationInterceptor.mapper;
+
+import com.baomidou.mybatisplus.mapper.BaseMapper;
+import com.baomidou.mybatisplus.test.mysql.paginationInterceptor.entity.PageUser;
+
+public interface PageUserMapper extends BaseMapper<PageUser> {
+
+}
