+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/mysql/paginationInterceptor/service/PageUserService.java
@@ -1,0 +1,8 @@
+package com.baomidou.mybatisplus.test.mysql.paginationInterceptor.service;
+
+import com.baomidou.mybatisplus.service.IService;
+import com.baomidou.mybatisplus.test.mysql.paginationInterceptor.entity.PageUser;
+
+public interface PageUserService extends IService<PageUser> {
+
+}
