--- mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/plugins/optimisticLocker/mapper/TimestampVersionUserMapper.java
@@ -1,8 +1,0 @@
-package com.baomidou.mybatisplus.test.plugins.optimisticLocker.mapper;
-
-import com.baomidou.mybatisplus.mapper.BaseMapper;
-import com.baomidou.mybatisplus.test.plugins.optimisticLocker.entity.TimestampVersionUser;
-
-public interface TimestampVersionUserMapper extends BaseMapper<TimestampVersionUser> {
-
-}
