--- mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/plugins/paginationInterceptor/mapper/PageUserMapper.java
@@ -1,8 +1,0 @@
-package com.baomidou.mybatisplus.test.plugins.paginationInterceptor.mapper;
-
-import com.baomidou.mybatisplus.mapper.BaseMapper;
-import com.baomidou.mybatisplus.test.plugins.paginationInterceptor.entity.PageUser;
-
-public interface PageUserMapper extends BaseMapper<PageUser> {
-
-}
