--- mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/test/plugins/paginationInterceptor/service/PageUserService.java
@@ -1,8 +1,0 @@
-package com.baomidou.mybatisplus.test.plugins.paginationInterceptor.service;
-
-import com.baomidou.mybatisplus.service.IService;
-import com.baomidou.mybatisplus.test.plugins.paginationInterceptor.entity.PageUser;
-
-public interface PageUserService extends IService<PageUser> {
-
-}
