+++ mybatis-plus-core/src/test/resources/plugins/optimisticLockerInterceptor.xml
@@ -5,7 +5,7 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
            http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
-	<context:component-scan base-package="com.baomidou.mybatisplus.test.plugins.optimisticLocker" />
+	<context:component-scan base-package="com.baomidou.mybatisplus.test.h2.optimisticLocker" />
 
 	<import resource="classpath:h2/spring-jdbc.xml" />
 
@@ -13,7 +13,7 @@
 	<bean id="sqlSessionFactory"
 		class="com.baomidou.mybatisplus.spring.MybatisSqlSessionFactoryBean">
 		<property name="dataSource" ref="dataSource" />
-		<property name="typeAliasesPackage" value="com.baomidou.mybatisplus.test.plugins.optimisticLocker.entity" />
+		<property name="typeAliasesPackage" value="com.baomidou.mybatisplus.test.h2.optimisticLocker.entity" />
 		<property name="plugins">
 			<array>
 				<bean class="com.baomidou.mybatisplus.plugins.OptimisticLockerInterceptor"/>
@@ -28,7 +28,7 @@
 	</bean>
 
 	<bean id="mapperScannerConfigurer" class="org.mybatis.spring.mapper.MapperScannerConfigurer">
-		<property name="basePackage" value="com.baomidou.mybatisplus.test.plugins.optimisticLocker.mapper" />
+		<property name="basePackage" value="com.baomidou.mybatisplus.test.h2.optimisticLocker.mapper" />
 	</bean>
 
 	<bean id="globalConfig" class="com.baomidou.mybatisplus.entity.GlobalConfiguration">
